Resolution independe camera rendering
Window resizing requires updating the camera to avoid stretching and
tearing. This commit stores in the camera the current viewport width
and height and check on update if window size has changed to update
the camera accordingly. Also the rendering system is now sequential to
give more control over the rendering layers ordering.